+++ app/views/study.py
@@ -6,12 +6,12 @@ from htmlmin import minify
 from flask import Blueprint
 from flask import render_template, jsonify, redirect, request, url_for, abort, flash
 from flask_login import login_required, current_user
-from app import db
+from app import db, csrf
 from app.models import Device
 from app.models import LessonType, Lesson, Video
 from app.decorators import permission_required
 from app.utils import get_mac_address_from_ip
-from app.utils import y_system_api_request
+from app.utils import y_system_api_request, verify_data_keys
 from app.utils2 import add_user_log
 
 
@@ -141,15 +141,19 @@ def video(id):
     '''study.video(id)'''
     video = Video.query.get_or_404(id)
     if not current_user.can_study(lesson=video.lesson):
-        flash('请先完成本课程的前序内容！', category='warning')
+        flash('无法研修当前课程：{}'.format(video.lesson.name), category='warning')
         return redirect(url_for('study.{}'.format(video.lesson.type.view_point)))
-    if not current_user.punched(video=video):
-        add_user_log(
-            user=current_user._get_current_object(),
-            event='视频研修：{}'.format(video.name),
-            category='study'
-        )
-        db.session.commit()
+    if video.lesson.type.name in ['VB', 'Y-GRE', 'Y-GRE AW']:
+        data = y_system_api_request(api='lesson-access', token_data={
+            'user_id': current_user.id,
+            'lesson': video.lesson.name,
+        })
+        if data is None:
+            flash('网络通信故障', category='error')
+            return redirect(url_for('study.{}'.format(video.lesson.type.view_point)))
+        if verify_data_keys(data=data, keys=['error']):
+            flash('无法研修当前课程：{}'.format(data.get('error')), category='error')
+            return redirect(url_for('study.{}'.format(video.lesson.type.view_point)))
     return minify(render_template(
         'study/video.html',
         video=video
@@ -161,21 +165,36 @@ def video(id):
 @permission_required('研修')
 def punch(id):
     '''study.punch(id)'''
+    csrf.protect()
     video = Video.query.get_or_404(id)
-    if not current_user.can('研修{}'.format(video.lesson.type.name)) or \
-        not current_user.can_study(lesson=video.lesson):
+    if not current_user.can_play(video=video):
         abort(403)
     if request.json is None:
         abort(500)
+    if not current_user.punched(video=video):
+        add_user_log(
+            user=current_user._get_current_object(),
+            event='视频研修：{}'.format(video.name),
+            category='study'
+        )
     current_user.punch(video=video, play_time=request.json.get('play_time'))
     db.session.commit()
     if video.lesson.type.name in ['VB', 'Y-GRE', 'Y-GRE AW']:
-        y_system_api_request(api='punch', token_data={
-            'user_id': current_user.id,
-            'section': video.name if video.lesson.type.name == 'VB' \
-                else '{} 视频研修'.format(video.lesson.name),
-            'progress': current_user.video_progress(video=video) \
-                if video.lesson.type.name == 'VB' \
-                else current_user.lesson_progress(lesson=video.lesson),
-        })
-    return jsonify(current_user.video_punch(video=video).to_json())
+        # synchronize study progress with Y-System
+        punch = current_user.get_punch(video=video)
+        if punch.sync_required:
+            data = y_system_api_request(api='punch', token_data={
+                'user_id': current_user.id,
+                'section': video.section,
+            })
+            if verify_data_keys(data=data, keys=['success']):
+                punch.set_synchronized()
+                add_user_log(
+                    user=current_user._get_current_object(),
+                    event='同步研修进度至Y-System：{}'.format(video.section),
+                    category='study'
+                )
+                db.session.commit()
+    return jsonify({
+        'progress': current_user.video_progress(video=video),
+    })