+++ config.py
@@ -23,7 +23,11 @@ class Config:
     # SSL
     SSL_DISABLE = True
 
+    # CSRF
+    WTF_CSRF_CHECK_DEFAULT = False
+
     # Security
+    SECRET_KEY = os.getenv('YVOD_SECRET_KEY')
     AUTH_TOKEN_SECRET_KEY = os.getenv('YVOD_AUTH_TOKEN_SECRET_KEY')
     AUTH_REMEMBER_LOGIN = True
 
@@ -49,8 +53,16 @@ class Config:
     # Video Analytics
     VIDEO_ANALYTICS_ACCELERATING_FACTOR = 1.25 # speedup
     VIDEO_ANALYTICS_GRANULARITY = 100 # milliseconds (0.1 seconds)
-    VIDEO_ANALYTICS_UPDATE_PUNCH_INTERVAL = 15 * 1000 # milliseconds (15 seconds)
+    VIDEO_ANALYTICS_UPDATE_PUNCH_INTERVAL = 60 * 1000 # milliseconds (1 minute)
     VIDEO_ANALYTICS_STATUS_EXPIRATION = 300 # seconds (5 minutes)
+
+    # Y-System
+    YSYS_URI = os.getenv('YVOD_YSYS_URL')
+
+    # Development
+    SYSTEM_OPERATOR_NAME = os.getenv('YVOD_SYSTEM_OPERATOR_NAME')
+    SERVER_SERIAL = os.getenv('YVOD_SERVER_SERIAL')
+    SERVER_MAC_ADDRESS = os.getenv('YVOD_SERVER_MAC_ADDRESS')
 
     # Version
     PYTHON_VERSION = '{0.major}.{0.minor}.{0.micro}'.format(sys.version_info)
@@ -71,23 +83,11 @@ class DevelopmentConfig(Config):
     ENV = 'development'
     DEBUG = True
 
-    # Security
-    SECRET_KEY = os.getenv('YVOD_DEV_SECRET_KEY')
-
     # Database
     SQLALCHEMY_DATABASE_URI = 'sqlite:///{}'.format(os.path.join(basedir, 'yvod-dev.sqlite'))
 
-    # Development
-    SYSTEM_OPERATOR_NAME = os.getenv('YVOD_DEV_SYSTEM_OPERATOR_NAME')
-    SYSTEM_OPERATOR_TOKEN = os.getenv('YVOD_DEV_SYSTEM_OPERATOR_TOKEN')
-    SERVER_SERIAL = os.getenv('YVOD_DEV_SERVER_SERIAL')
-    SERVER_MAC_ADDRESS = os.getenv('YVOD_DEV_SERVER_MAC_ADDRESS')
-
     # HLS
     HLS_ENABLE = False
-
-    # Y-System
-    YSYS_URI = os.getenv('YVOD_DEV_YSYS_URL')
 
 
 class ProductionConfig(Config):
@@ -95,23 +95,45 @@ class ProductionConfig(Config):
 
     ENV = 'production'
 
-    # Security
-    SECRET_KEY = os.getenv('YVOD_PROD_SECRET_KEY')
-
     # Database
     SQLALCHEMY_DATABASE_URI = os.getenv('YVOD_PROD_DATABASE_URL')
 
-    # Development
-    SYSTEM_OPERATOR_NAME = os.getenv('YVOD_PROD_SYSTEM_OPERATOR_NAME')
-    SYSTEM_OPERATOR_TOKEN = os.getenv('YVOD_PROD_SYSTEM_OPERATOR_TOKEN')
-    SERVER_SERIAL = os.getenv('YVOD_PROD_SERVER_SERIAL')
-    SERVER_MAC_ADDRESS = os.getenv('YVOD_PROD_SERVER_MAC_ADDRESS')
+    # Mail
+    MAIL_SERVER = 'smtp.exmail.qq.com'
+    MAIL_PORT = 587
+    MAIL_USE_TSL = True
+    MAIL_USERNAME = os.getenv('YVOD_MAIL_USERNAME')
+    MAIL_PASSWORD = os.getenv('YVOD_MAIL_PASSWORD')
+    MAIL_SUBJECT_PREFIX = os.getenv('YVOD_MAIL_SUBJECT_PREFIX')
+    MAIL_SENDER = os.getenv('YVOD_MAIL_SENDER')
+    SYSTEM_OPERATOR_MAIL = os.getenv('YVOD_SYSTEM_OPERATOR_MAIL')
 
     # HLS
     HLS_ENABLE = True
 
-    # Y-System
-    YSYS_URI = os.getenv('YVOD_PROD_YSYS_URL')
+    @classmethod
+    def init_app(cls, app):
+        '''ProductionConfig.init_app(cls, app)'''
+        Config.init_app(app)
+        # email errors to the system operator
+        import logging
+        from logging.handlers import SMTPHandler
+        credentials = None
+        secure = None
+        if getattr(cls, 'MAIL_USERNAME', None) is not None:
+            credentials = (cls.MAIL_USERNAME, cls.MAIL_PASSWORD)
+            if getattr(cls, 'MAIL_USE_TSL', None):
+                secure = ()
+        mail_handler = SMTPHandler(
+            mailhost=(cls.MAIL_SERVER, cls.MAIL_PORT),
+            fromaddr=cls.MAIL_SENDER,
+            toaddrs=[cls.SYSTEM_OPERATOR_MAIL],
+            subject='{} Application Error'.format(cls.MAIL_SUBJECT_PREFIX),
+            credentials=credentials,
+            secure=secure
+        )
+        mail_handler.setLevel(logging.ERROR)
+        app.logger.addHandler(mail_handler)
 
 
 config = {